+++ app/Models/Token.php
@@ -49,6 +49,7 @@ class Token extends Model
         self::IMPERSONATION => 24 * 3600,
         self::EMAIL_APPROVE_WITHDRAWAL => 3600,
         self::WALLET_CONNECT => 60,
+        self::WALLET_CONNECT_CONFIRM => 1 * 3600,
         self::WALLET_LOGIN => 60,
     ];
 
@@ -75,6 +76,8 @@ class Token extends Model
     public const EMAIL_APPROVE_WITHDRAWAL = 'email-approve-withdrawal';
 
     public const WALLET_CONNECT = 'wallet-connect';
+
+    public const WALLET_CONNECT_CONFIRM = 'wallet-connect-confirm';
 
     public const WALLET_LOGIN = 'wallet-login';