+++ app/Http/Controllers/Manager/AuthController.php
@@ -28,6 +28,8 @@ use Adshares\Adserver\Mail\UserConfirmed;
 use Adshares\Adserver\Mail\UserEmailActivate;
 use Adshares\Adserver\Mail\UserEmailChangeConfirm1Old;
 use Adshares\Adserver\Mail\UserEmailChangeConfirm2New;
+use Adshares\Adserver\Mail\UserPasswordChange;
+use Adshares\Adserver\Mail\UserPasswordChangeConfirm;
 use Adshares\Adserver\Models\Config;
 use Adshares\Adserver\Models\RefLink;
 use Adshares\Adserver\Models\Token;
@@ -46,16 +48,18 @@ use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Mail;
 use Illuminate\Support\Facades\Validator;
 use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
-use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\HttpKernel\Exception\UnauthorizedHttpException;
 use Symfony\Component\HttpKernel\Exception\UnprocessableEntityHttpException;
 
 class AuthController extends Controller
 {
+    private const ERROR_MESSAGE_INVALID_TOKEN = 'Invalid or outdated token';
+
     private ExchangeRateReader $exchangeRateReader;
 
     public function __construct(ExchangeRateReader $exchangeRateReader)
@@ -241,7 +245,7 @@ class AuthController extends Controller
         if (false === $token = Token::check($token, null, Token::EMAIL_CHANGE_STEP_1)) {
             DB::commit();
 
-            return self::json([], Response::HTTP_FORBIDDEN, ['message' => 'Invalid or outdated token']);
+            return self::json([], Response::HTTP_FORBIDDEN, ['message' => self::ERROR_MESSAGE_INVALID_TOKEN]);
         }
         $user = User::findOrFail($token['user_id']);
         if (User::withTrashed()->where('email', $token['payload']['email'])->count()) {
@@ -270,7 +274,7 @@ class AuthController extends Controller
         if (false === $token = Token::check($token, null, Token::EMAIL_CHANGE_STEP_2)) {
             DB::commit();
 
-            return self::json([], Response::HTTP_FORBIDDEN, ['message' => 'Invalid or outdated token']);
+            return self::json([], Response::HTTP_FORBIDDEN, ['message' => self::ERROR_MESSAGE_INVALID_TOKEN]);
         }
         $user = User::findOrFail($token['user_id']);
         if (User::withTrashed()->where('email', $token['payload']['email'])->count()) {
@@ -462,27 +466,38 @@ MSG;
             throw new UnauthorizedHttpException('', 'Required authenticated access or token authentication');
         }
 
-        if (!$request->has('user.password_new')) {
-            throw new BadRequestHttpException('Field `user.password_new` is required.');
-        }
-        $this->validateRequestObject($request, 'user', User::$rules);
+        $this->validateRequestObject($request, 'user', ['password_new' => 'required|min:8']);
 
         DB::beginTransaction();
         if (Auth::check()) {
             $user = Auth::user();
-            $token_authorization = false;
+            $tokenAuthorization = false;
         } else {
             if (false === $token = Token::check($request->input('user.token'), null, Token::PASSWORD_RECOVERY)) {
                 DB::rollBack();
-
                 throw new UnprocessableEntityHttpException('Authentication token is invalid');
             }
             $user = User::findOrFail($token['user_id']);
-            $token_authorization = true;
+            $tokenAuthorization = true;
+        }
+
+        if (!$tokenAuthorization && null === $user->password) {
+            DB::rollBack();
+            if (null === $user->email || !$user->is_email_confirmed) {
+                throw new UnprocessableEntityHttpException('Email is not set');
+            }
+            Validator::make($request->all(), ['uri' => 'required'])->validate();
+            $confirmToken = Token::generate(
+                Token::PASSWORD_CHANGE,
+                $user,
+                ['password' => Hash::make($request->input('user.password_new'))]
+            );
+            Mail::to($user)->queue(new UserPasswordChangeConfirm($confirmToken->uuid, $request->input('uri')));
+            return self::json([]);
         }
 
         if (
-            !$token_authorization &&
+            !$tokenAuthorization &&
             null !== $user->password &&
             (!$request->has('user.password_old') || !$user->validPassword($request->input('user.password_old')))
         ) {
@@ -496,6 +511,34 @@ MSG;
         }
 
         $user->password = $request->input('user.password_new');
+        $user->api_token = null;
+        $user->save();
+
+        if (null !== $user->email) {
+            Mail::to($user)->queue(new UserPasswordChange());
+        }
+
+        DB::commit();
+
+        return self::json($user->toArray());
+    }
+
+    public function confirmPasswordChange(string $token): JsonResponse
+    {
+        DB::beginTransaction();
+        if (false === $tokenData = Token::check($token, null, Token::PASSWORD_CHANGE)) {
+            DB::rollBack();
+            throw new UnprocessableEntityHttpException(self::ERROR_MESSAGE_INVALID_TOKEN);
+        }
+
+        $user = User::findOrFail($tokenData['user_id']);
+
+        if (null === $user->email || !$user->is_email_confirmed) {
+            DB::rollBack();
+            throw new UnprocessableEntityHttpException('Email is not set');
+        }
+
+        $user->setHashedPasswordAttribute($tokenData['payload']['password']);
         $user->api_token = null;
         $user->save();
         DB::commit();