+++ app/Models/Token.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Copyright (c) 2018-2021 Adshares sp. z o.o.
+ * Copyright (c) 2018-2022 Adshares sp. z o.o.
  *
  * This file is part of AdServer
  *
@@ -44,6 +44,7 @@ class Token extends Model
         self::EMAIL_ACTIVATE => 24 * 3600,
         self::EMAIL_CHANGE_STEP_1 => 24 * 3600,
         self::EMAIL_CHANGE_STEP_2 => 24 * 3600,
+        self::PASSWORD_CHANGE => 3600,
         self::PASSWORD_RECOVERY => 24 * 3600,
         self::IMPERSONATION => 24 * 3600,
         self::EMAIL_APPROVE_WITHDRAWAL => 3600,
@@ -56,6 +57,7 @@ class Token extends Model
         self::EMAIL_ACTIVATE => 5 * 60,
         self::EMAIL_CHANGE_STEP_1 => 5 * 60,
         self::EMAIL_CHANGE_STEP_2 => 5 * 60,
+        self::PASSWORD_CHANGE => 2 * 60,
         self::PASSWORD_RECOVERY => 2 * 60,
     ];
 
@@ -64,6 +66,8 @@ class Token extends Model
     public const EMAIL_CHANGE_STEP_1 = 'email-change-step1';
 
     public const EMAIL_CHANGE_STEP_2 = 'email-change-step2';
+
+    public const PASSWORD_CHANGE = 'password-change';
 
     public const PASSWORD_RECOVERY = 'password-recovery';