+++ CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Video ads support
 ### Changed
+- 2FA requirement when setting password
 - 2FA requirement when connecting the wallet
 
 ## [1.13.3] - 2022-02-10